+++ documentation/docs/changelog.md
@@ -3,6 +3,7 @@
 ## yyyy.mm.dd - version 1.4.0
 
 * Add a factory to create item instances from multiple sources
+* Add the Container model to map the backpack, bags, bank slots, etc
 * Add the id property to the Item model
 
 ## 2024.05.24 - version 1.3.0